+++ dlls/testdll.nim
@@ -26,6 +26,23 @@ withPragma { exportc, dynlib }:
     buffer = fmt"fastcall {a} {b} {c}"
     return cstring buffer
 
+  proc test_varargs(i: cint): cint {.cdecl, varargs.} =
+    {.emit: """
+      int sum = i;
+      va_list argp;
+      va_start(argp, i);
+
+      while(1) {
+        int arg_value = va_arg(argp, int);
+        sum += arg_value;
+        if (arg_value == 0)
+          break;
+      }
+      va_end(argp);
+
+      return sum;
+    """.}
+
   proc DllMain*(hinst: HINSTANCE, reason: DWORD, reserved: LPVOID): BOOL {.stdcall.} =
     if reason == DLL_PROCESS_ATTACH: NimMain()
     return TRUE