+++ changelog.md
@@ -1,3 +1,9 @@
+Version 1.2.0
+-------------
+* Fix cstring issue to work with winim 3.7.0.
+* Allows varargs pragma (#2).
+* Modify tests and examples to test varargs.
+
 Version 1.1.0
 -------------
 * Add memlib/rtlib submodule.